+++ app/__init__.py
@@ -1,0 +1,26 @@
+from flask import Flask
+from flask_login import LoginManager
+from flask_bootstrap import Bootstrap
+from flask_sqlalchemy import SQLAlchemy
+
+import os
+
+
+basedir = os.path.abspath(os.path.dirname(__file__))
+
+app = Flask(__name__)
+
+app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY') or 'Thisissupposedtobesecret!'
+app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('DATABASE_URL') or 'sqlite:///' + os.path.join(basedir, 'test.db')
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+
+bootstrap = Bootstrap(app)
+login_manager = LoginManager()
+login_manager.init_app(app)
+login_manager.login_view = 'login'
+
+db = SQLAlchemy(app)
+
+
+from app import old_app
+from app import routes
