+++ app/models.py
@@ -1,0 +1,28 @@
+from flask_login import UserMixin
+from app import db, login_manager
+
+
+@login_manager.user_loader
+def load_user(user_id):
+    return User.query.get(int(user_id))
+
+
+class Todo(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(200), nullable=False)
+    desc = db.Column(db.String(800), nullable=False)
+    date = db.Column(db.String(15), nullable=False)
+
+    def __repr__(self):
+        return '<Task %r>' % self.id
+
+
+class User(UserMixin, db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(15), unique=True)
+    email = db.Column(db.String(50), unique=True)
+    password = db.Column(db.String(80))
+
+    def __repr__(self):
+        return '<User %r>' % self.id
+
