+++ app/templates/result.html
@@ -18,12 +18,12 @@
     <script>
 
         var place=document.getElementById("wrapper")
-        var desc={{ sanswer.desc|tojson}}
-        var date1={{ sanswer.date|tojson}}
+        var desc={{ sanswer.desc }}
+        var date1={{ sanswer.date }}
         console.log(k);
         var date2=new Date()
         var val=((date2 - date1)/(1000*60*60))
-        var names={{ sanswer.title|tojson}}
+        var names={{ sanswer.title }}
         var data='<div id="countdown"><div class="well"><span id="hour" class="timer bg-success"></span><span id="min" class="timer bg-info"></span><span id="sec" class="timer bg-primary"></span></div></div>';
         var modal='<div class="modal fade container-fluid" id="exampleModalCenter" tabindex="-1" role="dialog" aria-labelledby="exampleModalCenterTitle" aria-hidden="true"><div class="modal-dialog modal-dialog-centered" role="document"><div class="modal-content">    <div class="modal-header"><h5 class="modal-title" id="exampleModalLongTitle">Modal title</h5><button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button></div><div class="modal-body">...</div><div class="modal-footer footer">'+data+'<button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button><button type="button" class="btn btn-primary">Remind Me</button></div></div></div></div>';
         var card='<div class="card mb-4 text-black bg-light" id ="'+names+'">'+
