+++ event_management.py
@@ -1,0 +1,5 @@
+from app import *
+
+
+if __name__ == "__main__":
+    app.run(debug=True)
